requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
aiofiles==23.2.1
python-dotenv==1.0.0
transformers==4.35.2
torch==2.1.1
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.3
PyPDF2==3.0.1
langchain==0.0.340
chromadb==0.4.18
faiss-cpu==1.7.4
tiktoken==0.5.1
openai==1.3.7
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
pymongo==4.6.0
celery==5.3.4
flower==2.0.1
prometheus-client==0.19.0
structlog==23.2.0